requirements.txt
pandas
apache-airflow-providers-amazon
faker 
boto3
apache-airflow-providers-snowflake